+++ kurtosis-module/impl/services/indexer.ts
@@ -1,19 +1,27 @@
-import { EnclaveContext, ServiceID, ContainerConfig, ContainerConfigBuilder, ServiceContext, PortSpec, PortProtocol } from "kurtosis-core-api-lib";
-import log = require("loglevel");
+import { FilesArtifactUUID, EnclaveContext, ServiceID, ContainerConfig, ContainerConfigBuilder, ServiceContext, PortSpec, PortProtocol } from "kurtosis-core-api-lib";
+import * as log from "loglevel";
 import { Result, ok, err } from "neverthrow";
 import { EXEC_COMMAND_SUCCESS_EXIT_CODE } from "../consts";
 import { ContainerConfigSupplier } from "../near_module";
 import { getPrivateAndPublicUrlsForPortId, ServiceUrl } from "../service_url";
+import * as path from "path";
 
 const SERVICE_ID: ServiceID = "indexer-node"
 const IMAGE: string = "kurtosistech/near-indexer-for-explorer:7510e7f";
-const RPC_PORT_NUM: number = 3030;
+const RPC_PRIVATE_PORT_NUM: number = 3030;
+const RPC_PUBLIC_PORT_NUM: number = 8332;
 const RPC_PORT_ID = "rpc";
-const RPC_PORT_SPEC = new PortSpec(RPC_PORT_NUM, PortProtocol.TCP);
+const RPC_PRIVATE_PORT_SPEC = new PortSpec(RPC_PRIVATE_PORT_NUM, PortProtocol.TCP);
+const RPC_PUBLIC_PORT_SPEC = new PortSpec(RPC_PUBLIC_PORT_NUM, PortProtocol.TCP);
 const RPC_PORT_PROTOCOL = "http";
-const GOSSIP_PORT_NUM: number = 24567;
+const GOSSIP_PRIVATE_PORT_NUM: number = 24567;
+const GOSSIP_PUBLIC_PORT_NUM: number = 8333;
 const GOSSIP_PORT_ID = "gossip";
-const GOSSIP_PORT_SPEC = new PortSpec(GOSSIP_PORT_NUM, PortProtocol.TCP);
+const GOSSIP_PRIVATE_PORT_SPEC = new PortSpec(GOSSIP_PRIVATE_PORT_NUM, PortProtocol.TCP);
+const GOSSIP_PUBLIC_PORT_SPEC = new PortSpec(GOSSIP_PUBLIC_PORT_NUM, PortProtocol.TCP);
+
+const LOCALNET_CONFIG_DIRPATH_ON_MODULE = "/static-files/near-configs/localnet"
+const NEAR_CONFIGS_DIRPATH_ON_INDEXER_CONTAINER = "/root/.near"
 
 const DATABASE_URL_ENVVAR = "DATABASE_URL";
 
@@ -42,9 +50,21 @@ export async function addIndexer(
     dbName: string,
 ): Promise<Result<IndexerInfo, Error>> {
     log.info(`Adding indexer service...`);
+
+    // Send the genesis file to Kurtosis
+    const uploadLocalnetConfigResult = await enclaveCtx.uploadFiles(LOCALNET_CONFIG_DIRPATH_ON_MODULE)
+    if (uploadLocalnetConfigResult.isErr()) {
+        return err(uploadLocalnetConfigResult.error)
+    }
+    const localnetConfigFilesArtifactUuid = uploadLocalnetConfigResult.value;
+
     const usedPorts: Map<string, PortSpec> = new Map();
-    usedPorts.set(RPC_PORT_ID, RPC_PORT_SPEC);
-    usedPorts.set(GOSSIP_PORT_ID, GOSSIP_PORT_SPEC);
+    usedPorts.set(RPC_PORT_ID, RPC_PRIVATE_PORT_SPEC);
+    usedPorts.set(GOSSIP_PORT_ID, GOSSIP_PRIVATE_PORT_SPEC);
+
+    const publicPorts: Map<string, PortSpec> = new Map();
+    publicPorts.set(RPC_PORT_ID, RPC_PUBLIC_PORT_SPEC);
+    publicPorts.set(GOSSIP_PORT_ID, GOSSIP_PUBLIC_PORT_SPEC);
 
     const envvars: Map<string, string> = new Map();
     envvars.set(
@@ -52,13 +72,20 @@ export async function addIndexer(
         `postgres://${dbUsername}:${dbUserPassword}@${dbPrivateUrl.ipAddress}:${dbPrivateUrl.portNumber}/${dbName}`
     )
 
+    const filesArtifactMounts = new Map<FilesArtifactUUID, string>();
+    filesArtifactMounts.set(localnetConfigFilesArtifactUuid, NEAR_CONFIGS_DIRPATH_ON_INDEXER_CONTAINER)
+
     const containerConfigSupplier: ContainerConfigSupplier = (ipAddr: string): Result<ContainerConfig, Error> => {
         const result: ContainerConfig = new ContainerConfigBuilder(
             IMAGE,
         ).withUsedPorts(
             usedPorts
+        ).withPublicPorts(
+            publicPorts,
         ).withEnvironmentVariableOverrides(
             envvars
+        ).withFiles(
+            filesArtifactMounts,
         ).build();
         return ok(result);
     }